pyproject.toml
[build-system]
# AVOID CHANGING REQUIRES: IT WILL BE UPDATED BY PYSCAFFOLD!
requires = ["setuptools>=46.1.0", "setuptools_scm[toml]>=5"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# For smarter version schemes and other configuration options,
# check out https://github.com/pypa/setuptools_scm
version_scheme = "no-guess-dev"

[project]
name = "palipedia"
authors = [{ name = "Erwin Jansen", email = "pokowaka@gmail.com" }]
license = { file = "LICENSE" }
version = "0.0.1"
classifiers = ["Topic :: Software Development"]


dependencies = ['absl-py', 'Unidecode', 'lxml']
